refactor: extract shared heading patterns to eliminate triple duplication
- Create headingPatterns.ts as single source of truth for section heading
  regex patterns, boundary regex, and detectTitle utility
- Fix PART heading missing from title extraction (was in boundary regex
  but not in pattern arrays — silent bug)
- sectionSplitter.ts and sectionExtractor.ts now import from shared module
- Collapse \r\n normalization in cleaner.ts to single regex pass
