+++ packages/ingestion/src/transformers/index.ts
@@ -3,3 +3,8 @@ export { createSplitter, splitDocuments } from "./splitter.js";
 export { enrichMetadata } from "./metadataEnricher.js";
 export { extractSections } from "./sectionExtractor.js";
 export { sectionAwareSplit } from "./sectionSplitter.js";
+export {
+  HEADING_PATTERNS,
+  HEADING_BOUNDARY,
+  detectTitle,
+} from "./headingPatterns.js";
